update readme with better explanations
