+++ cypress/integration/containers/SignUpPage.spec.js
@@ -1,9 +1,11 @@
 import faker from 'faker';
 import { signUpStub } from 'stubs/sessionStubs';
-import { FAIL_CASE } from 'cypressConstants';
+import { FAIL_CASE, inputTypes, validationTypes } from 'cypressConstants';
 import { testFields } from 'reusableTests';
 import routes from 'constants/routesPaths';
-import fields from '../../fields/signUpPageFields';
+
+const { INPUT } = inputTypes;
+const { PRESENCE, EMAIL, EQUALITY } = validationTypes;
 
 const email = faker.internet.email();
 const password = faker.internet.password();
@@ -41,6 +43,54 @@ describe('Sign Up Page', () => {
   });
 
   context('Form Validations', () => {
+    const fields = [
+      {
+        title: 'Email',
+        name: 'email',
+        inputType: INPUT,
+        errors: [
+          {
+            validationType: PRESENCE,
+            options: { customMessage: 'You must enter an email to continue' }
+          },
+          {
+            validationType: EMAIL,
+            options: { customMessage: 'You must enter a valid email' }
+          }
+        ]
+      },
+      {
+        title: 'Password',
+        name: 'password',
+        inputType: INPUT,
+        errors: [
+          {
+            validationType: PRESENCE,
+            options: { customMessage: 'You must enter a password to continue' }
+          }
+        ]
+      },
+      {
+        title: 'Password Confirmation',
+        name: 'passwordConfirmation',
+        inputType: INPUT,
+        errors: [
+          {
+            validationType: PRESENCE,
+            options: { customMessage: 'You must enter a password confirmation to continue' }
+          },
+          {
+            validationType: EQUALITY,
+            options: {
+              otherInput: 'password',
+              setup: () => {
+                cy.get('input[name=password]').type('password123');
+              }
+            }
+          }
+        ]
+      },
+    ];
     testFields(fields);
   });
